--- a/src/storage/mvcc/reader/reader.rs
+++ b/src/storage/mvcc/reader/reader.rs
@@ -58,6 +58,18 @@ impl<S: EngineSnapshot> SnapshotReader<S> {
         }
     }
 
+    pub fn new_scan_mode_with_ctx(
+        start_ts: TimeStamp,
+        snapshot: S,
+        scan_mode: ScanMode,
+        ctx: &Context,
+    ) -> Self {
+        SnapshotReader {
+            reader: MvccReader::new_with_ctx(snapshot, Some(scan_mode), ctx),
+            start_ts,
+        }
+    }
+
     #[inline(always)]
     pub fn get_txn_commit_record(&mut self, key: &Key) -> Result<TxnCommitRecord> {
         self.reader.get_txn_commit_record(key, self.start_ts)