--- a/src/storage/txn/commands/commit.rs
+++ b/src/storage/txn/commands/commit.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use tikv_kv::ScanMode;
 // #[PerformanceCriticalPath]
+use tikv_kv::ScanMode;
 use txn_types::Key;
 
 use crate::storage::{