--- a/src/storage/txn/commands/prewrite.rs
+++ b/src/storage/txn/commands/prewrite.rs
@@ -9,6 +9,7 @@
 use std::mem;
 
 use engine_traits::CF_WRITE;
+use itertools::Itertools;
 use kvproto::kvrpcpb::{
     AssertionLevel, ExtraOp, PrewriteRequestForUpdateTsConstraint,
     PrewriteRequestPessimisticAction::{self, *},
@@ -537,10 +538,19 @@ impl<K: PrewriteKind> Prewriter<K> {
         self.check_max_ts_synced(&snapshot)?;
 
         let mut txn = MvccTxn::new(self.start_ts, context.concurrency_manager);
-        let mut reader = ReaderWithStats::new(
-            SnapshotReader::new_with_ctx(self.start_ts, snapshot, &self.ctx),
-            context.statistics,
-        );
+        let snapshot_reader = if self.mutations.len() <= 1 {
+            SnapshotReader::new_with_ctx(self.start_ts, snapshot, &self.ctx)
+        } else {
+            let snapshot_reader = SnapshotReader::new_scan_mode_with_ctx(
+                self.start_ts,
+                snapshot,
+                tikv_kv::ScanMode::Forward,
+                &self.ctx,
+            );
+            self.mutations.sort_by(|a, b| a.key().cmp(&b.key()));
+            snapshot_reader
+        };
+        let mut reader = ReaderWithStats::new(snapshot_reader, context.statistics);
         // Set extra op here for getting the write record when check write conflict in
         // prewrite.
 
@@ -616,7 +626,6 @@ impl<K: PrewriteKind> Prewriter<K> {
 
         // If there are other errors, return other error prior to `AssertionFailed`.
         let mut assertion_failure = None;
-
         for m in mem::take(&mut self.mutations) {
             let pessimistic_action = m.pessimistic_action();
             let expected_for_update_ts = m.pessimistic_expected_for_update_ts();
@@ -885,6 +894,7 @@ trait MutationLock {
     fn pessimistic_action(&self) -> PrewriteRequestPessimisticAction;
     fn pessimistic_expected_for_update_ts(&self) -> Option<TimeStamp>;
     fn into_mutation(self) -> Mutation;
+    fn key(&self) -> &Key;
 }
 
 impl MutationLock for Mutation {
@@ -898,6 +908,10 @@ impl MutationLock for Mutation {
 
     fn into_mutation(self) -> Mutation {
         self
+    }
+
+    fn key(&self) -> &Key {
+        self.key()
     }
 }
 
@@ -925,6 +939,10 @@ impl MutationLock for PessimisticMutation {
 
     fn into_mutation(self) -> Mutation {
         self.mutation
+    }
+
+    fn key(&self) -> &Key {
+        self.mutation.key()
     }
 }