--- a/src/storage/txn/commands/prewrite.rs
+++ b/src/storage/txn/commands/prewrite.rs
@@ -13,7 +13,7 @@ use kvproto::kvrpcpb::{
     AssertionLevel, ExtraOp, PrewriteRequestForUpdateTsConstraint,
     PrewriteRequestPessimisticAction::{self, *},
 };
-use tikv_kv::SnapshotExt;
+use tikv_kv::{ScanMode, SnapshotExt};
 use txn_types::{
     insert_old_value_if_resolved, Key, Mutation, OldValues, TimeStamp, TxnExtra, Write, WriteType,
 };
@@ -537,20 +537,12 @@ impl<K: PrewriteKind> Prewriter<K> {
         self.check_max_ts_synced(&snapshot)?;
 
         let mut txn = MvccTxn::new(self.start_ts, context.concurrency_manager);
-        let snapshot_reader = if self.mutations.len() <= 1 {
-            SnapshotReader::new_with_ctx(self.start_ts, snapshot, &self.ctx)
-        } else {
+        let mut snapshot_reader = SnapshotReader::new_with_ctx(self.start_ts, snapshot, &self.ctx);
+
+        if self.mutations.len() > 1 {
             self.mutations.sort_by(|a, b| a.key().cmp(b.key()));
-            let mut snapshot_reader = SnapshotReader::new_scan_mode_with_ctx(
-                self.start_ts,
-                snapshot,
-                tikv_kv::ScanMode::Forward,
-                &self.ctx,
-            );
-            snapshot_reader
-                .reader
-                .set_range(self.mutations.first().map(|m| m.key().clone()), None);
-            snapshot_reader
+            snapshot_reader.set_scan_mode(ScanMode::Forward);
+            snapshot_reader.set_lower_bound(self.mutations.first().unwrap().key().clone());
         };
         let mut reader = ReaderWithStats::new(snapshot_reader, context.statistics);
         // Set extra op here for getting the write record when check write conflict in