<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed
-->

### What is changed and how it works?
<!--
Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#linking-issues.
-->

Issue Number: Close #18208

<!--
You could use "commit message" code block to add more description to the final commit message.
For more info, check https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md#format-of-the-commit-message.
-->
What's Changed:

```commit-message
When the write schedulers need to read multiple keys, it can be done using a cursor, which includes near-seek optimizations.
This PR leverages the cursor to minimize costly seek operations.
```

### Related changes

- [ ] PR to update `pingcap/docs`/`pingcap/docs-cn`:
- [ ] Need to cherry-pick to the release branch

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [x] Manual test (add detailed scripts or steps below)
- [ ] No code

Pipelined DML which resolve 2e8 locks
- nightly: 43m31s

![image](https://github.com/user-attachments/assets/22edbfa6-9916-4323-95e7-9daf24b6cff7)

- this PR: 36m51s

![image](https://github.com/user-attachments/assets/a677e39e-83ba-46fd-adf5-7cc4452f484e)


Range non-index update on sysbench table (50 continous keys)

- nightly: 4494tps

- this PR: 5624tps

![CPU](https://github.com/user-attachments/assets/ba014cb5-024c-4d2d-8b80-4282446dfff3)

Scheduler prewrite

![Scheduler prewrite](https://github.com/user-attachments/assets/b1fadd7e-2e38-4d10-9b2a-dfb6f0c5e6ef)

Scheduler commit

![Scheduler commit](https://github.com/user-attachments/assets/3f968ed8-25b9-451d-86cd-52533ed52eb0)

Sysbench bulk insert, the bottleneck is not in the scheduler side, so the result is close. However, the CPU usage in scheduler pool is quite different.

- nightly: 279223tps
- this PR: 284872tps

![image](https://github.com/user-attachments/assets/740404ef-2b04-457c-8112-babae0d3bedb)

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

### Release note
<!-- 
Compatibility change, improvement, bugfix, and new feature need a release note.

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

If you don't think this PR needs a release note then fill it with None.
If this PR will be picked to release branch, then a release note is probably required.
-->

```release-note
Optimize the write performance by using read cursor in transaction scheduler.
```